+++ lib/src/visitor/to_css_visitor.dart
@@ -633,7 +633,6 @@ class ToCSSVisitor extends VisitorBase{
 //  }
   }
 
-
   /// func visitor.visit distribuitor
   Function visitFtn(Node node) {
     if (node is Comment)    return this.visitComment;